refactor: MVC의 비즈니스 로직 Presenter로 이동
